all is working there is left to implement a listener to the delete X on the search box

+++ EvaluacionFinal7/app/src/main/java/com/exercises/sart1991/evaluacionfinal7/data/AppDataManager.java
@@ -111,7 +111,7 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
-    public List<Donor> getAllDonors(int donorId) {
+    public List<Donor> getAllDonors(long donorId) {
         return dbHelper.getAllDonors(donorId);
     }
